pyproject.toml
[project]
name = "apartment-search"
version = "0.1.0"
description = "Semantic search system for apartments using vector embeddings and multi-modal analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "orjson>=3.10.0",
    "google-generativeai>=0.8.5",
    "google-genai>=0.2.0",
    "elasticsearch>=8.11.0,<9.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.1.2",
    "python-multipart>=0.0.9",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
